# openapi.yaml
openapi: "3.0.3"
info:
  title: "Domain Authority & SEO Checker API"
  version: "1.0.0"
  description: |
    Instant domain authority scores, backlink metrics, SEO health,
    and spam analysis for agencies, SaaS platforms, and automated
    SEO workflows. Sub-100ms response times via edge infrastructure.
  contact:
    name: "API Support"
    email: "support@your-api.com"
  license:
    name: "Commercial"
servers:
  - url: "https://domain-seo-checker.p.rapidapi.com"
    description: "RapidAPI Production"
security:
  - RapidAPIKey: []
paths:
...
  /domain/spam-score:
    get:
      summary: "Get Spam Score"
      operationId: "getSpamScore"
      tags: ["Spam"]
      parameters:
        - name: domain
          in: query
          required: true
          description: "Target domain"
          schema:
            type: string
            example: "example.com"
      responses:
        '200':
          description: "Spam analysis returned successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpamScoreResponse'
              example:
                spam_score: 12
                risk_level: "low"
                manual_action_risk: false
                signals:
                  - signal: "thin_content"
                    triggered: false
                  - signal: "exact_match_anchors"
                    triggered: true
                  - signal: "low_quality_links"
                    triggered: false
        '400':
          description: "Invalid domain format"
        '429':
          description: "Rate limit exceeded"
...
components:
  securitySchemes:
    RapidAPIKey:
      type: apiKey
      in: header
      name: X-RapidAPI-Key
      description: "Your RapidAPI subscription key"
  schemas:
...
    SpamScoreResponse:
      type: object
      properties:
        spam_score:
          type: integer
          minimum: 0
          maximum: 100
        risk_level:
          type: string
          enum: [low, medium, high]
        manual_action_risk:
          type: boolean
        signals:
          type: array
          items:
            type: object
            properties:
              signal:
                type: string
              triggered:
                type: boolean
...